+++ gingko/gingko-domain-impl/src/main/java/com/qingbo/gingko/domain/inner/AccountLogServiceInner.java
@@ -3,9 +3,9 @@ package com.qingbo.gingko.domain.inner;
 import com.qingbo.gingko.common.result.Result;
 
 public interface AccountLogServiceInner {
-	Result<Boolean> deposit(Integer transactionId);
-	Result<Boolean> withdraw(Integer transactionId);
-	Result<Boolean> transfer(Integer transactionId);
-	Result<Boolean> freeze(Integer transactionId);
-	Result<Boolean> unfreeze(Integer transactionId);
+	Result<Boolean> deposit(Long transactionId);
+	Result<Boolean> withdraw(Long transactionId);
+	Result<Boolean> transfer(Long transactionId);
+	Result<Boolean> freeze(Long transactionId);
+	Result<Boolean> unfreeze(Long transactionId);
 }
