+++ gingko/gingko-domain-impl/src/main/java/com/qingbo/gingko/domain/inner/AccountServiceInner.java
@@ -3,9 +3,9 @@ package com.qingbo.gingko.domain.inner;
 import com.qingbo.gingko.common.result.Result;
 
 public interface AccountServiceInner {
-	Result<Boolean> deposit(Integer accountLogId);
-	Result<Boolean> withdraw(Integer accountLogId);
-	Result<Boolean> freeze(Integer accountLogId);
-	Result<Boolean> unfreeze(Integer accountLogId);
-	Result<Boolean> handle(Integer accountLogId);
+	Result<Boolean> deposit(Long accountLogId);
+	Result<Boolean> withdraw(Long accountLogId);
+	Result<Boolean> freeze(Long accountLogId);
+	Result<Boolean> unfreeze(Long accountLogId);
+	Result<Boolean> handle(Long accountLogId);
 }
