+++ gingko/gingko-repository/src/main/java/com/qingbo/gingko/repository/UserEnterpriseProfileRepository.java
@@ -3,5 +3,5 @@ package com.qingbo.gingko.repository;
 import com.qingbo.gingko.entity.UserEnterpriseProfile;
 
 public interface UserEnterpriseProfileRepository extends BaseRepository<UserEnterpriseProfile> {
-	UserEnterpriseProfile findByUserId(Integer userId);
+	UserEnterpriseProfile findByUserId(Long userId);
 }
