+++ gingko/gingko-web/src/main/java/com/qingbo/gingko/web/service/UserService.java
@@ -3,7 +3,7 @@ package com.qingbo.gingko.web.service;
 import com.qingbo.gingko.entity.User;
 
 public interface UserService {
-	Integer register(User user);
+	Long register(User user);
 	User getUser(String userName);
 	boolean[] login(String userName, String password);
 	void logout();
